feat: Implementa DTO de entrada para Veiculo e corrige serialização bidirecional
- Adiciona VeiculoInputDTO para gerenciar a associação de Veiculos via 'vendedorId'.
- Aplica @JsonManagedReference/@JsonBackReference nas entidades Veiculo e Vendedor para evitar o StackOverflowError na serialização JSON.

+++ src/main/java/com/example/minha_api_vendas/service/VeiculoService.java
@@ -1,8 +1,12 @@
 package com.example.minha_api_vendas.service;
 
+import com.example.minha_api_vendas.DTO.VeiculoDTO;
 import com.example.minha_api_vendas.model.Veiculo;
+import com.example.minha_api_vendas.model.Vendedor;
 import com.example.minha_api_vendas.repository.VeiculoRepository;
+import org.springframework.beans.InvalidPropertyException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.context.config.ConfigDataResourceNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,8 +18,22 @@ public class VeiculoService {
     @Autowired
     private VeiculoRepository _veiculoRepository;
 
-    public Veiculo Salvar(Veiculo veiculo)
-    {
+    @Autowired
+    private VendedorService _vendedorService;
+
+    public Veiculo Salvar(VeiculoDTO veiculoDTO) throws Exception {
+        Vendedor vendedor = _vendedorService.BuscarVendedorPorId(veiculoDTO.getVendedorId())
+                .orElseThrow(() -> new Exception("Vendedor não encontrado"));
+
+
+        Veiculo veiculo = new Veiculo();
+        veiculo.setAno(veiculoDTO.getAno());
+        veiculo.setMarca(veiculoDTO.getMarca());
+        veiculo.setModelo(veiculoDTO.getModelo());
+        veiculo.setPreco(veiculoDTO.getPreco());
+        veiculo.setVendido(veiculoDTO.getVendido());
+
+        veiculo.setVendedor(vendedor);
         return _veiculoRepository.save(veiculo);
     }
 
@@ -47,8 +65,12 @@ public class VeiculoService {
 
     public boolean DeletarVeiculoPorId(Long id)
     {
-        _veiculoRepository.deleteById(id);
-        return true;
+        if (_veiculoRepository.existsById(id))
+        {
+            _veiculoRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
 }
